Delete old crud.py
